# openapi.yaml
openapi: 3.1.0
info:
  title: ROSPlus Web-IDE API
  version: 0.2.0
  description: >-
    REST API for the ROSPlus Web-IDE backend. Provides workspace management,

    build scheduling, node execution, graph topology, and metrics.


    WebSocket endpoints are documented separately in design.md (Boundary E).


    Version 0.2 describes the local development service. Native transport
    metrics may be null until measured. Workspaces execute trusted local code
    and are not security sandboxes.
  contact:
    name: ROSPlus
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://{host}:{port}
    description: Production (HTTPS required)
    variables:
      host:
        default: localhost
      port:
        default: "8443"
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/workspace/create:
    post:
      tags:
        - Workspace
      summary: Create a new workspace
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "201":
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Workspace with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Maximum workspace limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/workspace:
    get:
      tags:
        - Workspace
      summary: List persisted workspaces
      operationId: listWorkspaces
      responses:
        "200":
          description: Workspace list
          content:
            application/json:
              schema:
                type: object
                required:
                  - workspaces
                properties:
                  workspaces:
                    type: array
                    items:
                      $ref: "#/components/schemas/Workspace"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API token authentication
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Workspace not found
        code:
          type: integer
          example: 404
        details:
          type: string
    CreateWorkspaceRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Human-readable workspace name
          example: my_first_robot
          minLength: 1
          maxLength: 64
          pattern: ^[a-zA-Z0-9_-]+$
        template:
          type: string
          description: Project template to use
          enum:
            - empty
            - talker_listener
            - camera
          default: empty
    Workspace:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        path:
          type: string
          description: Filesystem path on the server
        status:
          type: string
          enum:
            - idle
            - building
            - running
            - error
        created:
          type: string
          format: date-time
        last_activity:
          type: string
          format: date-time
        node_count:
          type: integer
...